pyproject.toml
[project]
name = "fit-agent"
version = "0.1.0"
description = "AI fitness and nutrition agent with meal photo analysis"
requires-python = ">=3.10"
authors = [
    { name = "Sourav Modak", email = "s.modak@uni-hohenheim.de" }
]
license = { text = "MIT" }

dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "openai>=1.0.0",
    "streamlit>=1.38.0",
    "plotly>=5.24.0",
    "pillow>=10.4.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=21.0",
    "mlflow>=2.16.0",
    "rich>=13.0.0",
    "whylogs",
    "prometheus-client",
    "groq"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]